refactor(core): organize domain models into Models directory with dedicated files
