+++ Sources/LogBird/Models/LBLogEvent.swift
