+++ Sources/LogBird/Models/LBLogMessage.swift
